+++ addons/binding/org.openhab.binding.sonyaudio/META-INF/MANIFEST.MF
@@ -7,8 +7,6 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-SymbolicName: org.openhab.binding.sonyaudio;singleton:=true
 Bundle-Vendor: openHAB
 Bundle-Version: 2.5.0.qualifier
-Export-Package: org.openhab.binding.sonyaudio,
- org.openhab.binding.sonyaudio.handler
 Import-Package: com.google.gson,
  com.google.gson.reflect,
  org.apache.commons.lang,
