+++ addons/binding/org.openhab.binding.sonyaudio/src/main/java/org/openhab/binding/sonyaudio/internal/SonyAudioHandlerFactory.java
@@ -19,14 +19,13 @@ import org.eclipse.smarthome.core.thing.binding.BaseThingHandlerFactory;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
 import org.eclipse.smarthome.core.thing.binding.ThingHandlerFactory;
 import org.eclipse.smarthome.io.net.http.WebSocketFactory;
-import org.openhab.binding.sonyaudio.SonyAudioBindingConstants;
-import org.openhab.binding.sonyaudio.handler.HtCt800Handler;
-import org.openhab.binding.sonyaudio.handler.HtMt500Handler;
-import org.openhab.binding.sonyaudio.handler.HtSt5000Handler;
-import org.openhab.binding.sonyaudio.handler.HtZ9fHandler;
-import org.openhab.binding.sonyaudio.handler.HtZf9Handler;
-import org.openhab.binding.sonyaudio.handler.SrsZr5Handler;
-import org.openhab.binding.sonyaudio.handler.StrDn1080Handler;
+import org.openhab.binding.sonyaudio.internal.handler.HtCt800Handler;
+import org.openhab.binding.sonyaudio.internal.handler.HtMt500Handler;
+import org.openhab.binding.sonyaudio.internal.handler.HtSt5000Handler;
+import org.openhab.binding.sonyaudio.internal.handler.HtZ9fHandler;
+import org.openhab.binding.sonyaudio.internal.handler.HtZf9Handler;
+import org.openhab.binding.sonyaudio.internal.handler.SrsZr5Handler;
+import org.openhab.binding.sonyaudio.internal.handler.StrDn1080Handler;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
